backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
PyPDF2